core: add stub implementation for REE-FS NV counter
Provide a placeholder for the non-volatile counter when using REE-FS.
This avoids enabling CFG_INSECURE while allowing REE-FS to function.

Note: REE-FS by itself, cannot provide a monotonic counter with security
guarantees comparable to RPMB. This implementation does not prevent
rollback and should be replaced if a reliable and hardware backed counter
is provided.

Signed-off-by: Theodore Karatapanis <tkaratapanis@census-labs.com>